Fix template error when generating basic auth
This addresses the issue referenced here:
https://github.com/gobuffalo/buffalo-auth/issues/54

The problem is that we have .plush files that themselves generate
.plush files.  There are two required fixes:

By escaping the <% tags in the plush files themselves we can
seperate the tags that should be evaluated at the time of
the genny.Transform vs those that should be evaluted in the
context of the webapp.

Once the plush file escaping issue is dealt with we still have
the problem that plushgen strips the .plush. extensions from
the file names (leaving new.html instead of new.plush.html).
I've added a new transform in the pipeline to rename ".html"
files back to ".plush.html" so that they will have the
appropriate file extention in the destination project.